Initialize  - Change Application ID  - New structure
