+++ .metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 8661d8aecd626f7f57ccbcb735553edc05a2e713
+  revision: 7a4c33425ddd78c54aba07d86f3f9a4a0051769b
   channel: stable
 
 project_type: app
