+++ android/app/src/main/kotlin/io/github/tdh8316/thanks/MainActivity.kt
@@ -1,0 +1,13 @@
+package io.github.tdh8316.thanks
+
+import android.os.Bundle
+
+import io.flutter.app.FlutterActivity
+import io.flutter.plugins.GeneratedPluginRegistrant
+
+class MainActivity: FlutterActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    GeneratedPluginRegistrant.registerWith(this)
+  }
+}
