--- lib/core/db/graph_provider.dart
@@ -1,5 +1,0 @@
-Future<List<int>> graphProvider() async {
-  // String stringData = await readFile(".json");
-
-  return <int>[];
-}
