+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:thanks/core/loader.dart';
+import 'package:thanks/screens/splash.dart';
 
 void main() => runApp(App());
 
@@ -27,7 +27,7 @@ class App extends StatelessWidget {
       title: "All that Thanks",
       theme: ThemeData(
         primarySwatch: Colors.lightBlue,
-        fontFamily: "나눔바른펜",
+        fontFamily: "나눔스퀘어라운드",
       ),
 
       // Display the splash screen
