+++ lib/screens/splash.dart
@@ -1,0 +1,68 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:thanks/core/i18n.dart';
+import 'package:thanks/screens/index.dart';
+
+class Splash extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() => _SplashState();
+}
+
+class _SplashState extends State<Splash> {
+  @override
+  void initState() {
+    super.initState();
+    Future.delayed(Duration(), () async {
+      await initializeStrings(Localizations.localeOf(context).toString());
+    });
+    Future.delayed(Duration(seconds: 2)).then(
+      (_) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(
+            builder: (context) => Index(),
+          ),
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      resizeToAvoidBottomPadding: false,
+      backgroundColor: Colors.white,
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          Center(
+            child: Center(
+              child: Text(
+                "All that thanks",
+                style: TextStyle(
+                  fontSize: 40,
+                  fontWeight: FontWeight.bold,
+                  foreground: Paint()
+                    ..shader = LinearGradient(
+                      end: Alignment.topRight,
+                      colors: <Color>[Colors.amber[400], Color(0xff8921aa)],
+                    ).createShader(
+                      Rect.fromLTWH(0.0, 0.0, 200.0, 70.0),
+                    ),
+                ),
+              ),
+            ),
+          ),
+          Padding(padding: EdgeInsets.only(top: 8)),
+          Center(
+            child: Text(
+              "Made with 💕 by 2019 DreamHigh",
+              style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
