+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: thanks
-description: All that Thanks
+description: What are you thankful for today?
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -23,21 +23,12 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  intl: ^0.15.7
-  animated_text_kit: ^1.3.0
-  intro_views_flutter: ^1.0.0
-  path_provider: ^0.5.0+1
-  encrypt: ^2.0.0
-  zefyr:
-    git:
-      url: https://github.com/memspace/zefyr.git
-      ref: flutter-master
-      path: packages/zefyr
 
-  flip_box_bar:
-  date_utils:
-  tuple:
+  path_provider:
+  intl:
 
 dev_dependencies:
   flutter_test:
@@ -60,10 +51,10 @@ flutter:
     - i18n/ko_KR.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.io/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware.
 
   # For details regarding adding assets from package dependencies, see
-  # https://flutter.io/assets-and-images/#from-packages
+  # https://flutter.dev/assets-and-images/#from-packages
 
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
@@ -71,9 +62,9 @@ flutter:
   # list giving the asset and other descriptors for the font. For
 
   fonts:
-    - family: 나눔바른펜
+    - family: 나눔스퀘어라운드
       fonts:
-        - asset: fonts/nanumbarunpen.ttf
+        - asset: fonts/NanumSquareRoundR.ttf
 
   # For details regarding fonts from package dependencies,
-  # see https://flutter.io/custom-fonts/#from-packages
+  # see https://flutter.dev/custom-fonts/#from-packages
